+++ src/cue/maestrocuehandler.cpp
@@ -5,7 +5,7 @@
 namespace PixelMaestro {
 
 	uint8_t* MaestroCueHandler::remove_show() {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::MaestroCueHandler,
 			(uint8_t)Action::RemoveShow
 		);
@@ -14,7 +14,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* MaestroCueHandler::set_brightness(uint8_t brightness) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::MaestroCueHandler,
 			(uint8_t)Action::SetBrightness
 		);
@@ -24,7 +24,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* MaestroCueHandler::set_show() {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::MaestroCueHandler,
 			(uint8_t)Action::SetShow
 		);
@@ -33,7 +33,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* MaestroCueHandler::set_timer(uint16_t interval) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::MaestroCueHandler,
 			(uint8_t)Action::SetTimer
 		);
@@ -43,7 +43,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* MaestroCueHandler::start() {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::MaestroCueHandler,
 			(uint8_t)Action::Start
 		);
@@ -52,7 +52,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* MaestroCueHandler::stop() {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::MaestroCueHandler,
 			(uint8_t)Action::Stop
 		);
@@ -61,7 +61,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* MaestroCueHandler::sync(const uint32_t new_time) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::MaestroCueHandler,
 			(uint8_t)Action::Sync
 		);
