+++ src/cue/sectioncuehandler.cpp
@@ -3,7 +3,7 @@
 
 namespace PixelMaestro {
 	uint8_t* SectionCueHandler::remove_animation(uint8_t section_num, uint8_t layer_num, bool clear_pixels) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::RemoveAnimation,
 			section_num,
@@ -15,7 +15,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* SectionCueHandler::remove_canvas(uint8_t section_num, uint8_t layer_num) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::RemoveCanvas,
 			section_num,
@@ -26,7 +26,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* SectionCueHandler::remove_layer(uint8_t section_num, uint8_t layer_num) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::RemoveLayer,
 			section_num,
@@ -38,7 +38,7 @@ namespace PixelMaestro {
 
 	uint8_t* SectionCueHandler::set_animation(uint8_t section_num, uint8_t layer_num, AnimationType animation_type, bool preserve_settings) {
 
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::SetAnimation,
 			section_num,
@@ -51,7 +51,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* SectionCueHandler::set_brightness(uint8_t section_num, uint8_t layer_num, uint8_t brightness) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::SetBrightness,
 			section_num,
@@ -63,7 +63,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* SectionCueHandler::set_canvas(uint8_t section_num, uint8_t layer_num, uint16_t num_frames) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::SetCanvas,
 			section_num,
@@ -75,7 +75,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* SectionCueHandler::set_dimensions(uint8_t section_num, uint8_t layer_num, uint16_t x, uint16_t y) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::SetDimensions,
 			section_num,
@@ -89,7 +89,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* SectionCueHandler::set_layer(uint8_t section_num, uint8_t layer_num, Colors::MixMode mix_mode, uint8_t alpha) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::SetLayer,
 			section_num,
@@ -102,7 +102,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* SectionCueHandler::set_mirror(uint8_t section_num, uint8_t layer_num, bool x, bool y) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::SetMirror,
 			section_num,
@@ -115,7 +115,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* SectionCueHandler::set_offset(uint8_t section_num, uint8_t layer_num, int16_t x, int16_t y) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::SetOffset,
 			section_num,
@@ -129,7 +129,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* SectionCueHandler::set_scroll(uint8_t section_num, uint8_t layer_num, uint16_t x, uint16_t y, bool reverse_x, bool reverse_y) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::SectionCueHandler,
 			(uint8_t)Action::SetScroll,
 			section_num,
